news-analyzer-python/requirements.txt
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.6.0
nltk==3.9.1
hdbscan==0.8.33
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pyyaml==6.0.2
tqdm==4.66.5
